separate file for amt p&m tests
